+++ src/proton/types/message_content/invoice.cr
@@ -17,7 +17,7 @@ module Proton::Types
     property currency : ::String
 
     # Product total price in the minimal quantity of the currency
-    property total_amount : ::Int64
+    property total_amount : ::Int32
 
     # Unique invoice bot start_parameter. To share an invoice use the URL https://t.me/{bot_username}?start={start_parameter}
     property start_parameter : ::String
@@ -29,12 +29,12 @@ module Proton::Types
     property need_shipping_address : ::Bool
 
     # The identifier of the message with the receipt, after the product has been purchased
-    property receipt_message_id : ::Int64
+    property receipt_message_id : ::Int32
 
     # Product photo; may be null
     property photo : Proton::Types::Photo? = nil
 
-    def initialize(@title : ::String, @description : ::String, @currency : ::String, @total_amount : ::Int64, @start_parameter : ::String, @is_test : ::Bool, @need_shipping_address : ::Bool, @receipt_message_id : ::Int64, @photo : Proton::Types::Photo? = nil)
+    def initialize(@title : ::String, @description : ::String, @currency : ::String, @total_amount : ::Int32, @start_parameter : ::String, @is_test : ::Bool, @need_shipping_address : ::Bool, @receipt_message_id : ::Int32, @photo : Proton::Types::Photo? = nil)
     end
   end
 end
