+++ src/proton/types/update/message_content_opened.cr
@@ -10,12 +10,12 @@ module Proton::Types
     include JSON::Serializable
 
     # Chat identifier
-    property chat_id : ::Int64
+    property chat_id : ::Int32
 
     # Message identifier
-    property message_id : ::Int64
+    property message_id : ::Int32
 
-    def initialize(@chat_id : ::Int64, @message_id : ::Int64)
+    def initialize(@chat_id : ::Int32, @message_id : ::Int32)
     end
   end
 end
